+++ app/users/admin/__init__.py
@@ -1,0 +1,1 @@
+from admin import UserAdmin
